Refactored getter and setters. Added Rarity enum (future use). Added Player armor slot getter and setters. Unit Tests updated

+++ CombatLooter/Classes/Interface/IBeing.cs
@@ -50,7 +50,7 @@ namespace CombatLooter.Classes.Interface
         /// Method to change Armor value
         /// </summary>
         /// <param name="amount">Amount of armor to change</param>
-        void ChangeArmor(int amount);
+        void ChangeArmor(double amount);
 
         /// <summary>
         /// Method to change Stamina value
